rename overlay screen to screenlock
